message-board-render/requirements.txt
flask==2.0.3
psycopg2-binary==2.9.3
python-dotenv==0.19.2